requirement.txt
azure-identity 
azure-kusto-data 
azure-kusto-ingest